Added unit tests for user input and results
Passing using $stdout and $stdin global variables as default parameters for the object allows us to both enter input when we run script manually, as well as perform unit testing on simulated user input. 
This is done by passing creating a StringIO object with the desired input as it's contents and passing this into the RPN_Calculator object as input.